Only Update Every X Hours
Added in some code that checks when a player's most recent games were
polled, and if it was within a certain number of hours (currently 24)
that player is skipped. Eventually I plan to add a list of players who
are updated regardless of if they were updated in the last 24 hours.